backend/requirements.txt
fastapi
uvicorn[standard]
langchain-groq
langchain-core
python-dotenv
pydantic
reportlab